% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCfertDHS.R
\name{APCfertDHS}
\alias{APCfertDHS}
\title{Compute Lexis fertility surface from a DHS survey and plot using APCplot}
\usage{
APCfertDHS(
  survey,
  fpath = "",
  byVar = NULL,
  subsetVar = NULL,
  subsetVal = NA,
  ...
)
}
\arguments{
\item{survey}{Survey name (the .dta extension is not required)}

\item{fpath}{Path to the directory containing the survey data file}

\item{byVar}{Categorical variable defining sub-groups for separate plots}

\item{subsetVar}{Variable identifying a sub-set of respondents to be plotted}

\item{subsetVal}{Code identifying the respondents to be kept}

\item{...}{Parameters to be passed to function \code{APCplot}}
}
\value{
Data frame of birth rates indexed by age, period and cohort of mother
}
\description{
\code{APCfertDHS} calculates single-year age-period-cohort birth rates from the
birth histories reported in a Demographic and Health Survey. Women's birth
cohort is defined by their age at interview and period is measured backward
from the time of interview but an option exists to label the axes in calendar
years.
}
\details{
Users must first register on the Demographic and Health Surveys
\href{https://dhsprogram.com}{web site} and request access to the survey(s) of
interest. After downloading the relevant zipped flat (i.e. wide) Stata
individual recode data files (??IR??DT.ZIP), they should extract the Stata
.dta files to their working directory before running \code{APCfertDHS}. The
function returns the Lexis data frame that it passes to APCplot, so users can
save it to use to produce further plots by calling \code{APCplot} directly.
}
\examples{
# Use a flat (i.e. wide) individual recode data file from a synthetic DHS 
# survey (ZZIR62FL). (To save space, the file only contains a few variables).
 
data(DHSurvey)

APCfertDHS(ZZIR62FL, byVar = "v190")
}
