% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCmortHMD.R
\name{APCmortHMD}
\alias{APCmortHMD}
\title{Compute Lexis mortality surface from HMD data and plot using APCplot}
\usage{
APCmortHMD(
  user,
  password,
  country_id,
  base_year = 1922L,
  length_yrs = 100L,
  ...
)
}
\arguments{
\item{user}{The email address used to create an HMD account.}

\item{password}{The password for that HMD account.}

\item{country_id}{Country code listed in the data section of the HMD website.}

\item{base_year}{Initial birth cohort for which rates are required.}

\item{length_yrs}{Number of APC categories for which rates are required.}

\item{...}{Parameters to be passed to function \code{APCplot}}
}
\value{
Data frame with death rates by sex indexed by age, period and cohort.
}
\description{
\code{APCmortHMD} is a function to download and reorganise mortality data from
the Human Mortality Database (HMD) for passing to APCplot.
}
\details{
Users must first register on the Human Mortality Database
\href{https://mortality.org}{web site} in order to download the data (N. B. If you
registered before June 2022, you need to re-register).

Note that not all countries have a century-long run of data and the plotting
function has not yet been generalized for use with axes of differing length
(e.g. 50 years of data on ages 0 to 99).
}
\examples{
## Load synthetic HMD data files as the real ones are password protected
data(Deaths)
data(Exposures)

## For the avoidance of doubt, Atlantis isn't real and neither are these data
id <- "ATLANTIS"
Lexis <- APCmortHMD(user = yourID, password = yourPW, country_id = id, 
base_year = 1921)
}
