% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCplot-package.R
\docType{package}
\name{APCplot-package}
\alias{APCplot-package}
\title{APCplot: Triangular Age-Period-Cohort surfaces}
\description{
Plot age-period-cohort surfaces (rates and rate ratios) on a grid of equilateral triangles
}
\keyword{internal}
