% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{DHSurvey}
\alias{DHSurvey}
\title{Demographic and Health Survey individual recode data imported using haven}
\format{
{ A data frame with 8348 rows and 321 columns:}
}
\source{
These are synthetic data
}
\usage{
data(DHSurvey)
}
\description{
Data frame for input to \code{setupDHSdata}. These are a subset of the
variables in a synthetic DHS data set (ZZIR61DT.DTA) that can be downloaded
from the \emph{Demographic and Health Survey} web site. These data are open
access but users are required to register and apply for access to real DHS
datasets.
}
\references{
Demographic and Health Survey: \url{dhsprogram.com}
}
\keyword{datasets}
