% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Exposures}
\alias{Exposures}
\title{Exposure data imported from the Human Mortality Database using HMDHFDplus}
\format{
{ A data frame with 1000s of rows and 7 columns:}
\itemize{
\item{\code{Year}}:  Calendar year
\item{\code{Age}}:  Age in years
\item{\code{Cohort}}:  Birth year
\item{\code{Female}}:  Count of years of exposure of females
\item{\code{Men}}:  Count of years of exposure of males
\item{\code{Total}}:  Count of years of exposure of both sexes
\item{\code{OpenInterval}}:  \code{TRUE} for the open-ended age interval
}
}
\source{
These are synthetic data
}
\usage{
data(Exposures)
}
\description{
Data frame for input to \code{setupHMDdata}. This data frame mirrors the
datasets that can be downloaded from \emph{Human Mortality Database} data by
registered users.
}
\references{
Human Mortality Database: \url{mortality.org}
}
\keyword{datasets}
