% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Lexis}
\alias{Lexis}
\title{Lexis data frame for plotting by APCplot}
\format{
{ A data frame with 10,200 rows and 9 columns:}
\itemize{
\item{\code{Year}}:  {Calendar year - (unused, copied from the HMD files)}
\item{\code{Age}}:  {Age in years - (unused, copied from the HMD files)}
\item{\code{Cohort}}:  {Birth year - (unused, copied from the HMD files)}
\item{\code{Women}}:  {Women's death rates calculated as Deaths/Exposures}
\item{\code{Men}}:  {Men's death rates calculated as Deaths/Exposures}
\item{\code{Rates}}:  {Total death rates calculated as Deaths/Exposures}
\item{\code{age}}:  {Age index \code{<- Age}}
\item{\code{coh}}:  {Cohort index \code{<- Cohort - Base_year}}
\item{\code{per}}:  {Reversed period index \code{<- Base_year+99 - Year}}
}
}
\source{
These are synthetic data
}
\usage{
data(Lexis)
}
\description{
\code{Lexis} data frame for input to \code{APCplot}. This data frame mirrors
the datasets created from \emph{Human Mortality Database} data by
\code{setupHMDdata} and is used by the examples in the \code{APCplot} help
file. It also illustrative of the structure of Lexis data frames for anyone
preparing them from scratch. N.B. In this dataset time is measured forward
from a base date and the period axis has been reversed. If time were being
measured backward (e.g. as years before a date of fieldwork of 0), the cohort
axis would need to be reversed instead.
}
\references{
Human Mortality Database: \url{mortality.org}
}
\keyword{datasets}
