% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LifeCourses.R
\name{LifeCourses}
\alias{LifeCourses}
\title{Plots of life courses on a triangular Lexis diagram}
\usage{
LifeCourses(
  YrB,
  AgeD,
  censored = FALSE,
  Events = NULL,
  EvType = "",
  Events2 = NULL,
  EvType2 = "",
  Events3 = NULL,
  EvType3 = "",
  base_year = 0L,
  survey_year = 0L,
  base_age = 0L,
  length_yrs = 100L,
  exact_data = FALSE,
  plot_title = "Example"
)
}
\arguments{
\item{YrB}{Vector of years of birth (either calendar years or years ago).}

\item{AgeD}{Vector of ages at death or censoring in years}

\item{censored}{Vector indicating whether individuals died or were censored}

\item{Events}{List of vectors of ages at event (e.g. giving birth). NULL if none.}

\item{EvType}{Name describing the event for use in the key e.g. "Births"}

\item{Events2}{List of vectors of ages at second type of event. NULL if none.}

\item{EvType2}{Name describing the event for use in the key e.g."Marriage"}

\item{Events3}{List of vectors of ages at third type of event. NULL if none.}

\item{EvType3}{Name describing the event for use in the key e.g."Divorce"}

\item{base_year}{Initial birth cohort}

\item{survey_year}{Year in which retrospective data were collected}

\item{base_age}{Youngest age supplying data}

\item{length_yrs}{Number of cohorts required}

\item{exact_data}{Flags the use of exact ages/dates, rather than durations in
completed years}

\item{plot_title}{Title}
}
\value{
R graphics plot that can be exported to multiple graphics formats.
}
\description{
\code{LifeCourses} draws the life courses of one or more individuals on a
Lexis plot represented as a grid of equilateral triangles. The vectors of
years of birth and ages at death and lists of othe life events should all
refer to the same number of individuals (i.e. be the same length).
}
\examples{
## Initialise life course histories for 4 individuals
YearB <- c(1925, 1938, 1962, 2001)
AgeD <- c(70, 80, 50, NA)
evType <- "Birth of child"
events <- list(NULL, c(17,26), c(32), NULL)
evType2 <- "Marriage"
events2 <- list(c(22, 40), c(16), c(25), NULL)
evType3 <- "Marital dissolution"
events3 <- list(c(28), NULL, NULL, NULL)
Censored <- c(F, F, T, F)

## Draw lexis diagram with cohort and period measured in calendar years
LifeCourses(YearB, AgeD, Events = events, EvType = evType,  
                Events2 = events2, EvType2 = evType2, Events3 = events3, 
                EvType3 = evType3, censored = Censored, survey_year = 2020)

## Redraw, indexing cohort and period as years before the end of observation
YearB <- c(95, 82, 58, 19)
LifeCourses(YearB, AgeD, Events = events, EvType = evType,  
                Events2 = events2, EvType2 = evType2, Events3 = events3, 
                EvType3 = evType3, censored = Censored)
}
