% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCplot-package.R
\docType{package}
\name{APCplot-package}
\alias{APCplot-package}
\title{APCplot: Triangular Age-Period-Cohort surfaces}
\description{
Plot age-period-cohort surfaces (rates and rate ratios) on a grid of equilateral triangles using Martin R Smith's Ternary package. An ancillary function is provided to set up datasets from the Human Mortality Database for use by APCplot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BugBunny/APCplot}
}

}
\keyword{internal}
