% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCplot-package.R
\docType{package}
\name{APCplot-package}
\alias{APCplot-package}
\title{APCplot: Triangular Age-Period-Cohort surfaces}
\description{
Plot either age-period-cohort surfaces (rates and rate ratios) or individuals' life courses on a Lexis grid of equilateral triangles using Martin R Smith's Ternary package. Ancillary functions are provided to set up datasets containing rates from the Human Mortality Database and from Demographic and Health Survey birth histories for plotting by APCplot.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BugBunny/APCplot}
}

}
\keyword{internal}
