% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCplot.R
\name{APCplot}
\alias{APCplot}
\title{Plot age-period-cohort Lexis surfaces}
\usage{
APCplot(
  Lexis,
  base_year = 0L,
  length_yrs = 100L,
  label_interval = 10L,
  contour_plot = FALSE,
  sex_specific = TRUE,
  sex_differences = FALSE,
  change_in_rates = FALSE,
  log_rates = TRUE
)
}
\arguments{
\item{base_year}{Initial birth cohort}

\item{label_interval}{Interval in years between axis labels}

\item{contour_plot}{Draw a contour plot}

\item{sex_specific}{Plot men and women separately}

\item{sex_differences}{Plot the men:women rate ratios}

\item{change_in_rates}{Plot the ratios of successive rates}

\item{log_rates}{Plot the log rates or rate ratios}

\item{length_years}{Number of cohorts required}
}
\value{
R graphics plot that can be exported to multiple graphics formats.
}
\description{
\code{APCplot} produces age-period-cohort plots of Lexis surfaces (either rates or
rate ratios) on a grid of equilateral triangles.
}
\details{
The Lexis surfaces can comprise rates or trends in rates (ratios of rates for
successive periods) and either be logged or not. \code{APCplot} can produce plots
for Men and Women side-by-side, a single plot for both sexes, or plots of the
Men's:Women's rate ratios. Each of these types of plot can be drawn
either for the individual APC triangles or as a contour plot.

The function is implemented by torturing Martin Smith's \link{Ternary} package
into drawing APC surfaces.
}
\examples{
APCplot(Lexis, base_year = 1922)
APCplot(Lexis, base_year = 1922, contour_plot = TRUE, sex_specific = FALSE)
APCplot(Lexis, base_year = 1922, sex_differences = TRUE, log_rates = FALSE)
APCplot(Lexis, base_year = 1922, change_in_rates = TRUE)
}
