% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCplot.R
\name{APCplot}
\alias{APCplot}
\title{Plot age-period-cohort Lexis surfaces}
\usage{
APCplot(
  Lexis,
  base_year = 0L,
  survey_year = 0L,
  base_age = 0L,
  length_yrs = 100L,
  contour_plot = FALSE,
  group_specific = TRUE,
  total = NULL,
  group_ratios = FALSE,
  period_ratios = FALSE,
  log_rates = TRUE,
  pct_trim = 0.5
)
}
\arguments{
\item{Lexis}{Data frame with columns for age, per, coh and each set of rates.
The first 3 columns index the coordinates of the rates contained in the other
columns. Store the rates for the total population in a column named "Rates".
\link{APCmortHMD} prepares and plots Lexis mortality surfaces using HMD data
\link{APCfertDHS} prepares and plots Lexis fertility surfaces from a DHS survey}

\item{base_year}{Initial birth cohort}

\item{survey_year}{Year when retrospective data were collected}

\item{base_age}{Youngest age supplying data}

\item{length_yrs}{Number of cohorts required}

\item{contour_plot}{Draw a contour plot}

\item{group_specific}{Plot sub-groups (e.g. men and women) separately}

\item{total}{Add a plot for the total population to plots for sub-groups}

\item{group_ratios}{Plot the ratio of the rates in a pair of sub-groups}

\item{period_ratios}{Plot the ratios of successive rates}

\item{log_rates}{Plot the logged rates or rate ratios}

\item{pct_trim}{Percentage of rates to trim from each extreme of the range}
}
\value{
R graphics plot that can be exported to multiple graphics formats.
}
\description{
\code{APCplot} produces age-period-cohort plots of Lexis surfaces (either rates or
rate ratios) on a grid of equilateral triangles.
}
\details{
The Lexis surfaces can comprise rates or trends in rates (ratios of rates for
successive periods) and either be logged or not. \code{APCplot} can produce plots
for several sub-groups on a grid (e.g. men and women), a single plot of the
full data, or plots of the ratio of the rates in a pair of sub-groups. Each
of these types of plot can be drawn either for the individual APC triangles
or as a contour plot.

The function is implemented by torturing Martin Smith's \link{Ternary} package
into drawing APC surfaces.
}
\examples{
## Load a Lexis data frame of death rates by age, period and cohort
data(Lexis)
## Log rates for each sex
APCplot(Lexis, base_year = 1922)
## Contour plot of the log rates for the two sexes combined
APCplot(Lexis, base_year = 1922, contour_plot = TRUE, group_specific = FALSE)
## Ratios of the men's to the women's rates
APCplot(Lexis, base_year = 1922, group_ratios = TRUE, log_rates = FALSE)
## Rates of change in the rates for each sex
APCplot(Lexis, base_year = 1922, period_ratios = TRUE)
}
