% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Lexis}
\alias{Lexis}
\title{Lexis data frame for plotting by APCplot}
\format{
{ A data frame with 10,200 rows and 6 columns:}
\itemize{
\item{\code{age}}:  {Age index}
\item{\code{coh}}:  {Cohort index \code{<- Year_of_birth - Base_year}}
\item{\code{per}}:  {Reversed period index \code{<- Base_year+99 - Year}}#'
\item{\code{Women}}:  {Women's death rates calculated as Deaths/Exposures}
\item{\code{Men}}:  {Men's death rates calculated as Deaths/Exposures}
\item{\code{Rates}}:  {Total death rates calculated as Deaths/Exposures}
}
}
\source{
These are synthetic data
}
\usage{
data(Lexis)
}
\description{
\code{Lexis} data frame for input to \code{APCplot}. This data frame contains
100 years of data and mirrors the datasets created from \emph{Human Mortality
Database} data by \code{setupHMDdata}. It is used by the examples in the
\code{APCplot} help file. It also illustrative of the structure of Lexis data
frames for anyone preparing them from scratch. N.B. In this dataset time is
measured forward from a base date and the period axis has been reversed. If
time were being measured backward (e.g. as years before a date of fieldwork
of 0), the cohort axis would need to be reversed instead.
}
\keyword{datasets}
